fatto refresh su vecchio database compilato a mano, ed eliminate alcune migration superflue, dato il comando per effettuare il seeder e ricompilare il database da 0

+++ database/migrations/2022_05_13_145613_create_trains_table.php
@@ -22,7 +22,7 @@ class CreateTrainsTable extends Migration
             $table->time('departure_time');
             $table->time('arrival_time');
             $table->mediumInteger('train_code');
-            $table->integer('train_wagons');
+            $table->tinyInteger('train_wagons');
             $table->boolean('is_in_time');
             $table->boolean('was_deleted');
             $table->timestamps();
